+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 #### Filesystem
 - Added `Nickvision::Filesystem::ApplicationUserDirectory` enum
 - Added `Nickvision::Filesystem::UserDirectories::get()` function overload for `Nickvision::Filesystem::ApplicationUserDirectory` support
+#### Helpers
+- Added `Nickvision::Helpers::StringHelpers::normalizeForFilename()`
+- Added `Nickvision::Helpers::StringHelpers::replace()` character overload
 ### Fixes
 #### Update
 - Fixed an issue where `Nickvision::Update::Updater::windowsUpdate()` did not work correctly